requirements.txt
pandas 
numpy
pymongo[srv]==3.12
pymongo
dotenv
certifi
scikit-learn
dill
pyaml
# -e .